Moved TestWriter
