+++ src/com/audiophile/t2m/music/MelodyTrack.java
@@ -1,7 +1,6 @@
 package com.audiophile.t2m.music;
 
 import com.audiophile.t2m.text.Sentence;
-import com.audiophile.t2m.io.TestWriter;
 
 import javax.sound.midi.*;
